backend/requirements.txt
Flask>=2.0
Flask-SQLAlchemy>=2.5
psycopg2-binary>=2.9
Flask-Migrate>=3.0
python-dotenv>=0.15 # Keep this for loading .env
Flask-CORS>=4.0 # Allow frontend to connect to backend
Flask-JWT-Extended>=4.0 # JWT token-based authentication
bcrypt>=4.0 # Password hashing
